Change of variable from int to float type

+++ Service/src/main/java/me/upp/arquis/greenwaste/services/materials/routes/MaterialCreate.java
@@ -15,7 +15,7 @@ public class MaterialCreate extends RouteWrapper {
     @Override
     public Object handle(final Request request, final Response response){
         final String name = request.params(":name");
-        final int cost = Integer.parseInt(request.params(":cost"));
+        final float cost = Float.parseFloat(request.params(":cost"));
         final String insert = new LInsert()
                 .table("material")
                 .values(
